matroska: separate EBML parsing into an independant module
This will ease the parsing of other EBML-based formats. It also checks
the EBML doctype, so matroska will not probe positively on any EBML
file.
